Switched from simple linked list to Hash Table in the Memory Management Module for greater speed

+++ Simulator/Simulator/Simulator/include/MemoryManagement.h
@@ -5,7 +5,7 @@
 
 
 // number of memory allocations that are currently not freed
-uint32_t memory_manage_memory_blocks();
+size_t memory_manage_memory_blocks();
 
 // number of bytes currently not freed
 size_t memory_manage_memory_size();
